+++ entry/node/main.ts
@@ -1,2 +1,3 @@
 export * from "!/src/core/CTraderConnection";
 export * from "!/src/core/CTraderConnectionParameters";
+export * from "!/src/core/types";